backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
diffusers==0.23.0
transformers==4.35.0
accelerate==0.24.1
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.26.4
insightface==0.7.3
onnxruntime==1.16.1
reportlab==4.0.7
redis==5.0.1
rq==1.16.2
prometheus-client==0.20.0
safetensors==0.4.3
gunicorn==21.2.0
